+++ subscription/mongodb/native/blocking/src/main/java/org/occurrent/subscription/mongodb/nativedriver/blocking/BlockingSubscriptionForMongoDB.java
@@ -64,9 +64,11 @@ import static java.util.Objects.requireNonNull;
 
 /**
  * This is a subscription that uses the "native" MongoDB Java driver (sync) to listen to changes from the event store.
- * This Subscription doesn't maintain the subscription position, you need to store itin order to continue the stream
+ * This Subscription doesn't maintain the subscription position, you need to store it in order to continue the stream
  * from where it's left off on application restart/crash etc. You can do this yourself or use a
- * <a href="https://occurrent.org/documentation#blocking-subscription-position-storage">subscription position storage implementation</a>.
+ * <a href="https://occurrent.org/documentation#blocking-subscription-position-storage">subscription position storage implementation</a>
+ * or use the {@code BlockingSubscriptionWithAutomaticPositionPersistence} utility from the {@code org.occurrent:subscription-util-blocking-automatic-position-persistence}
+ * module.
  */
 public class BlockingSubscriptionForMongoDB implements PositionAwareBlockingSubscription {
     private static final Logger log = LoggerFactory.getLogger(BlockingSubscriptionForMongoDB.class);
@@ -204,7 +206,9 @@ public class BlockingSubscriptionForMongoDB implements PositionAwareBlockingSubs
 
     @Override
     public SubscriptionPosition globalSubscriptionPosition() {
-        BsonTimestamp currentOperationTime = MongoDBCommons.getServerOperationTime(database.runCommand(new Document("hostInfo", 1)));
+        // Note that we increase the "increment" by 1 in order to not clash with an existing event in the event store.
+        // This is so that we can avoid duplicates in certain rare cases when replaying events.
+        BsonTimestamp currentOperationTime = MongoDBCommons.getServerOperationTime(database.runCommand(new Document("hostInfo", 1)), 1);
         return new MongoDBOperationTimeBasedSubscriptionPosition(currentOperationTime);
     }